+++ src/resources/imports/imports.service.ts
@@ -1,13 +1,131 @@
 // src/imports/imports.service.ts
-import { Injectable } from '@nestjs/common';
-import { InjectQueue } from '@nestjs/bullmq';
-import { Queue } from 'bullmq';
+import { Injectable, Logger } from '@nestjs/common';
+import * as fs from 'fs';
+import * as path from 'path';
+import { parse } from 'fast-csv';
+import { PrismaService } from '../prisma/prisma.service';
+import { mapClientCsvToIntermediate } from './utils/client-csv-mapper.util';
+import { mapLabelCsvToIntermediate } from './utils/label-csv-mapper.util';
 
 @Injectable()
 export class ImportsService {
-  constructor(@InjectQueue('client-import') private clientImportQueue: Queue) {}
+  private readonly logger = new Logger(ImportsService.name);
+
+  constructor(private readonly prisma: PrismaService) {}
 
   async importClientsFromCsv(filePath: string) {
-    await this.clientImportQueue.add('client-import-job', { filePath });
+    if (!fs.existsSync(filePath)) {
+      throw new Error(`File not found at path: ${filePath}`);
+    }
+    const errorLogPath = path.join(
+      path.dirname(filePath),
+      `error_log_client_${Date.now()}.txt`,
+    );
+    const records = await this.readCsvFile(
+      filePath,
+      mapClientCsvToIntermediate,
+    );
+    if (records.length === 0) {
+      throw new Error(`No records to process in CSV file.`);
+    }
+    this.logger.log(`Total records loaded: ${records.length}.`);
+    for (const [i, record] of records.entries()) {
+      const { clientData, contractData, dmbData, addressData } = record;
+      try {
+        const country = await this.prisma.country.findFirst({
+          where: { name: addressData.countryName },
+        });
+        if (!country) {
+          throw new Error(`Country not found: ${addressData.countryName}`);
+        }
+        delete addressData.countryName;
+        await this.prisma.client.create({
+          data: {
+            ...clientData,
+            address: {
+              create: {
+                ...addressData,
+                country: { connect: { id: country.id } },
+              },
+            },
+            contract: { create: { ...contractData } },
+            dmb: { create: { ...dmbData } },
+          },
+        });
+      } catch (error) {
+        this.logError(errorLogPath, `Row ${i + 1}: ${error.message}`);
+      }
+    }
+    this.cleanUp(filePath, errorLogPath);
+    return { message: 'Client CSV processed successfully.' };
+  }
+
+  async importLabelsFromCsv(filePath: string) {
+    if (!fs.existsSync(filePath)) {
+      throw new Error(`File not found at path: ${filePath}`);
+    }
+    const errorLogPath = path.join(
+      path.dirname(filePath),
+      `error_log_label_${Date.now()}.txt`,
+    );
+    const records = await this.readCsvFile(filePath, mapLabelCsvToIntermediate);
+    if (records.length === 0) {
+      throw new Error(`No records to process in CSV file.`);
+    }
+    this.logger.log(`Total records loaded: ${records.length}.`);
+    for (const [i, record] of records.entries()) {
+      const { labelData, wp_id } = record;
+      try {
+        const client = await this.prisma.client.findUnique({
+          where: { wp_id: wp_id },
+        });
+        this.logger.log(
+          `wp_id: ${wp_id}, client: ${client ? client.id : 'No encontrado'}`,
+        );
+        if (!client) {
+          throw new Error(`Cliente con wp_id ${wp_id} no encontrado.`);
+        }
+        const existingLabel = await this.prisma.label.findUnique({
+          where: { name: labelData.name },
+        });
+        if (!existingLabel) {
+          await this.prisma.label.create({
+            data: { ...labelData, clientId: client.id },
+          });
+        } else {
+          throw new Error(`Label "${labelData.name}" already exists`);
+        }
+      } catch (error) {
+        this.logError(errorLogPath, `Row ${i + 1}: ${error.message}`);
+      }
+    }
+    this.cleanUp(filePath, errorLogPath);
+    return { message: 'Label CSV processed successfully.' };
+  }
+
+  private async readCsvFile<T>(
+    filePath: string,
+    mapFn: (row: any) => T,
+  ): Promise<T[]> {
+    return new Promise((resolve, reject) => {
+      const records: T[] = [];
+      fs.createReadStream(filePath)
+        .pipe(parse({ headers: true, ignoreEmpty: true, delimiter: ',' }))
+        .on('data', (row) => records.push(mapFn(row)))
+        .on('end', () => resolve(records))
+        .on('error', (error) => reject(error));
+    });
+  }
+
+  private logError(filePath: string, message: string) {
+    fs.appendFileSync(filePath, `${new Date().toISOString()} - ${message}\n`);
+  }
+
+  private cleanUp(filePath: string, errorLogPath: string) {
+    if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
+    this.logger.log(`Temporary file ${filePath} deleted successfully.`);
+    if (fs.existsSync(errorLogPath)) {
+      this.logger.log(`Error log saved at ${errorLogPath}`);
+    }
   }
 }
